joyaiSFT/requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
attrs==25.4.0
audioread==3.1.0
av==16.0.1
blessed==1.25.0
brotli==1.2.0
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cuda-bindings==13.1.1
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
docstring_parser==0.17.0
einops==0.8.1
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.20.0
fire==0.7.1
flash_attn @ file:///mnt/llm-data/users/wanghao277/some_pkgs/flash_attn-2.8.3+cu12torch2.8cxx11abiTRUE-cp311-cp311-linux_x86_64.whl#sha256=3d41b2fc55753faa7f45d6568ea73a96b96afb48b82994ab9b49bcbcb6c87588
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
gradio==5.45.0
gradio_client==1.13.0
graphviz==0.21
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.0
langchain-core==1.2.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.16.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
peft==0.14.0
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rouge-chinese==1.0.3
ruff==0.14.9
safehttpx==0.1.7
safetensors==0.5.3
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
shtab==1.8.0
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sse-starlette==3.0.4
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
torchviz==0.0.3
tqdm==4.67.1
transformers==4.51.3
triton==3.4.0
trl==0.9.6
typer==0.20.0
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.3
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0